gov_agents/requirements.txt
mcp
httpx
httpx-sse
pandas
requests
langchain==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
python-dotenv
fastapi
uvicorn
tabulate
google-generativeai
